% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_create_input_security_group}
\alias{medialive_create_input_security_group}
\title{Creates a Input Security Group}
\usage{
medialive_create_input_security_group(Tags, WhitelistRules)
}
\arguments{
\item{Tags}{A collection of key-value pairs.}

\item{WhitelistRules}{List of IPv4 CIDR addresses to whitelist}
}
\description{
Creates a Input Security Group
}
\section{Request syntax}{
\preformatted{medialive$create_input_security_group(
  Tags = list(
    "string"
  ),
  WhitelistRules = list(
    list(
      Cidr = "string"
    )
  )
)
}
}

\keyword{internal}
