% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_delete_reservation}
\alias{medialive_delete_reservation}
\title{Delete an expired reservation}
\usage{
medialive_delete_reservation(ReservationId)
}
\arguments{
\item{ReservationId}{[required] Unique reservation ID, e.g. '1234567'}
}
\description{
Delete an expired reservation.
}
\section{Request syntax}{
\preformatted{medialive$delete_reservation(
  ReservationId = "string"
)
}
}

\keyword{internal}
