% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_describe_schedule}
\alias{medialive_describe_schedule}
\title{Get a channel schedule}
\usage{
medialive_describe_schedule(ChannelId, MaxResults, NextToken)
}
\arguments{
\item{ChannelId}{[required] Id of the channel whose schedule is being updated.}

\item{MaxResults}{}

\item{NextToken}{}
}
\description{
Get a channel schedule
}
\section{Request syntax}{
\preformatted{medialive$describe_schedule(
  ChannelId = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
