% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialive_operations.R
\name{medialive_list_channels}
\alias{medialive_list_channels}
\title{Produces list of channels that have been created}
\usage{
medialive_list_channels(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{}

\item{NextToken}{}
}
\description{
Produces list of channels that have been created
}
\section{Request syntax}{
\preformatted{medialive$list_channels(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
